new CalculateAverageSessionsPerWeekUseCase + tests added missing resetSessionsForUser in SimpleHiitRepository - no implementation yet, need tests also added missing deleteForUser in SessionsDao - no tests yet

+++ app/src/main/java/fr/shining_cat/simplehiit/domain/models/UserStatistics.kt
@@ -6,5 +6,5 @@ data class UserStatistics(
     val averageSessionLengthSeconds: Int = 0,
     val longestStreakDays: Int = 0,
     val currentStreakDays: Int = 0,
-    val averageNumberOfSessionsPerWeek: Int = 0
+    val averageNumberOfSessionsPerWeek: Double = 0.00
 )
